Rename Room module from room3 to room
