--- room3/build.gradle.kts
@@ -1,71 +1,0 @@
-// Copyright 2026 The Android Open Source Project
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-import org.jetbrains.kotlin.gradle.dsl.JvmTarget
-
-plugins {
-    alias(libs.plugins.android.library)
-    alias(libs.plugins.ksp)
-}
-
-android {
-    namespace = "com.example.snippets.room3"
-    compileSdk = libs.versions.compileSdk.get().toInt()
-
-    defaultConfig {
-        minSdk = libs.versions.minSdk.get().toInt()
-        testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
-    }
-
-    buildTypes {
-        release {
-            isMinifyEnabled = false
-            proguardFiles(
-                getDefaultProguardFile("proguard-android-optimize.txt"),
-            )
-        }
-    }
-    compileOptions {
-        sourceCompatibility = JavaVersion.VERSION_17
-        targetCompatibility = JavaVersion.VERSION_17
-    }
-    kotlin {
-        compilerOptions {
-            jvmTarget = JvmTarget.fromTarget("17")
-        }
-    }
-}
-
-dependencies {
-    implementation(libs.androidx.core.ktx)
-    ksp(libs.androidx.room3.compiler)
-    implementation(libs.androidx.room3.runtime)
-    implementation(libs.androidx.room3.paging)
-    implementation(libs.androidx.room3.rxjava3)
-    implementation(libs.androidx.room3.guava)
-    implementation(libs.androidx.room3.livedata)
-    implementation(libs.androidx.room3.sqlite.wrapper)
-    implementation(libs.androidx.sqlite.framework)
-    implementation(libs.androidx.sqlite.bundled)
-    implementation(libs.androidx.sqlite.async)
-    implementation(libs.androidx.tracing)
-    implementation(libs.kotlinx.coroutines.android)
-    testImplementation(libs.junit)
-    androidTestImplementation(libs.androidx.test.ext.junit)
-    androidTestImplementation(libs.androidx.test.espresso.core)
-    androidTestImplementation(libs.androidx.test.core)
-    androidTestImplementation(libs.androidx.test.runner)
-    androidTestImplementation(libs.androidx.room3.testing)
-    androidTestImplementation(libs.kotlinx.coroutines.test)
-}
