--- room3/src/androidTest/java/com/example/snippets/room3/testingdb/AndroidTestingDbSnippets.kt
@@ -1,87 +1,0 @@
-/*
- * Copyright 2026 The Android Open Source Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.example.snippets.room3.testingdb
-
-import android.content.Context
-import androidx.room3.Database
-import androidx.room3.Entity
-import androidx.room3.PrimaryKey
-import androidx.room3.Dao
-import androidx.room3.Insert
-import androidx.room3.Query
-import androidx.room3.Room
-import androidx.room3.RoomDatabase
-import androidx.sqlite.driver.bundled.BundledSQLiteDriver
-import androidx.test.core.app.ApplicationProvider
-import androidx.test.ext.junit.runners.AndroidJUnit4
-import org.junit.After
-import org.junit.Assert.assertEquals
-import org.junit.Before
-import org.junit.Test
-import org.junit.runner.RunWith
-
-// [START android_room3_testing_db_android]
-import kotlinx.coroutines.test.runTest
-
-@RunWith(AndroidJUnit4::class)
-class SimpleEntityReadWriteTest {
-    private lateinit var userDao: UserDao
-    private lateinit var db: TestDatabase
-
-    @Before
-    fun createDb() {
-        val context = ApplicationProvider.getApplicationContext<Context>()
-        db = Room.inMemoryDatabaseBuilder<TestDatabase>(context)
-                .setDriver(BundledSQLiteDriver())
-                .build()
-        userDao = db.userDao()
-    }
-
-    @After
-    fun closeDb() {
-        db.close()
-    }
-
-    @Test
-    fun writeUserAndReadInList() = runTest {
-        val user = User(3, "george")
-        userDao.insert(user)
-        val byName = userDao.findUsersByName("george")
-        assertEquals(byName.single(), user)
-    }
-}
-// [END android_room3_testing_db_android]
-
-@Entity
-data class User(
-    @PrimaryKey val id: Int,
-    val name: String
-)
-
-@Dao
-interface UserDao {
-    @Insert
-    suspend fun insert(user: User)
-
-    @Query("SELECT * FROM User WHERE name = :name")
-    suspend fun findUsersByName(name: String): List<User>
-}
-
-@Database(entities = [User::class], version = 1)
-abstract class TestDatabase : RoomDatabase() {
-    abstract fun userDao(): UserDao
-}
